meta: change name to modbus-core
This is not a minimal modbus library. In fact, it has more code than necessary
in some cases because it needs to handle no_std.

This is a core modbus library. It is difficult to use on its own, but it can be
used to implement all kinds of modbus implementations -- async, network, serial,
etc.